feature: split system router on two, make new logs_router.py
