+++ gpt2giga/routers/__init__.py
@@ -1,8 +1,7 @@
 from .anthropic_router import router as anthropic_router
 from .api_router import router as api_router
-from .system_router import logs_api_router
-from .system_router import logs_router
-from .system_router import router as system_router
+from .logs_router import logs_api_router, logs_router
+from .system_router import system_router
 
 __all__ = [
     "anthropic_router",
